requirements.txt
asttokens==2.4.1
awscli==1.35.16
boto3==1.35.50
botocore==1.35.50
colorama==0.4.6
comm==0.2.2
cramjam==2.9.0
debugpy==1.8.7
decorator==5.1.1
docutils==0.16
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.5.0
fsspec==2024.10.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rsa==4.7.2
s3transfer==0.10.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
